refactor: move zoho-sign documents/ under lib/
Documents (registry, evaluation, assembly) are translation logic —
they map our proposal domain to Zoho's envelope domain.
