+++ src/shared/components/contract-status-panel/ui/agent-contract-view.tsx
@@ -12,7 +12,7 @@ import { Textarea } from '@/shared/components/ui/textarea'
 import { useInvalidation } from '@/shared/dal/client/use-invalidation'
 import { formatDate } from '@/shared/lib/formatters'
 import { cn } from '@/shared/lib/utils'
-import { ENVELOPE_DOCUMENT_LABELS } from '@/shared/services/providers/zoho-sign/documents/labels'
+import { ENVELOPE_DOCUMENT_LABELS } from '@/shared/services/providers/zoho-sign/lib/documents/labels'
 import { useTRPC } from '@/trpc/helpers'
 import { ACTION_TOOLTIPS } from '../constants/contract-statuses'
 import { useCreditCooldown } from '../hooks/use-credit-cooldown'
