+++ src/shared/services/contracts.service.ts
@@ -10,8 +10,8 @@ import { contractEventColumn, contractEventIdempotencyPolicy, shouldAutoApproveO
 import { countPdfPages } from '@/shared/lib/pdf/count-pdf-pages'
 import { pdfService } from '@/shared/services/pdf.service'
 import { ZOHO_SIGN_BASE_URL } from '@/shared/services/providers/zoho-sign/constants'
-import { assembleEnvelope } from '@/shared/services/providers/zoho-sign/documents/assemble-envelope'
-import { buildProposalContext } from '@/shared/services/providers/zoho-sign/documents/proposal-context'
+import { assembleEnvelope } from '@/shared/services/providers/zoho-sign/lib/documents/assemble-envelope'
+import { buildProposalContext } from '@/shared/services/providers/zoho-sign/lib/documents/proposal-context'
 import { buildSigningRequest } from '@/shared/services/providers/zoho-sign/lib/build-signing-request'
 import { dedupeSignerStatuses } from '@/shared/services/providers/zoho-sign/lib/dedupe-signer-statuses'
 import { getZohoAccessToken } from '@/shared/services/providers/zoho-sign/lib/get-access-token'
