+++ src/trpc/routers/proposals.router/contracts.router.ts
@@ -21,9 +21,9 @@ import { customerCrud } from '@/shared/entities/customers/dal/server/crud'
 import { proposalCrud } from '@/shared/entities/proposals/dal/server/crud'
 import { getFullView } from '@/shared/entities/proposals/dal/server/queries'
 import { contractService } from '@/shared/services/contracts.service'
-import { EnvelopeSelectionError, evaluateDocuments, validateEnvelopeSelection } from '@/shared/services/providers/zoho-sign/documents/evaluate'
-import { buildProposalContext } from '@/shared/services/providers/zoho-sign/documents/proposal-context'
-import { ENVELOPE_DOCUMENTS } from '@/shared/services/providers/zoho-sign/documents/registry'
+import { EnvelopeSelectionError, evaluateDocuments, validateEnvelopeSelection } from '@/shared/services/providers/zoho-sign/lib/documents/evaluate'
+import { buildProposalContext } from '@/shared/services/providers/zoho-sign/lib/documents/proposal-context'
+import { ENVELOPE_DOCUMENTS } from '@/shared/services/providers/zoho-sign/lib/documents/registry'
 
 import { createTRPCRouter } from '../../init'
 import { dalToTrpc } from '../../lib/dal-to-trpc'
